# apps/api/src/auth/dto/sign-up.dto.ts
import { ApiProperty, ApiSchema } from '@nestjs/swagger';
import {
  IsNotEmpty,
  IsString,
  IsStrongPassword,
  Length,
  MaxLength,
} from 'class-validator';

@ApiSchema({
  description: 'Sign-up credentials',
})
export class SignUpDto {
  @ApiProperty({
    example: '9JrHmsFZK8J-a',
    type: String,
  })
  @IsNotEmpty()
  @IsString()
  @IsStrongPassword({
    minLength: 12,
    minLowercase: 1,
    minNumbers: 1,
    minSymbols: 1,
    minUppercase: 1,
  })
  @MaxLength(127)
  password!: string;

  @ApiProperty({
    example: 'john-doe',
    type: String,
  })
  @IsNotEmpty()
  @IsString()
  @Length(3, 32)
  username!: string;
}