# apps/api/swagger.json
{"openapi":"3.0.0","paths":{"/books":{"post":{"operationId":"BooksController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookResponseDto"}}}},"400":{"description":"Invalid input"},"500":{"description":"Internal server error"}},"summary":"Create a new book","tags":["Books"]},"get":{"operationId":"BooksController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookResponseDto"}}}}},"500":{"description":"Internal server error"}},"summary":"List all the books","tags":["Books"]}},"/books/{bookId}":{"get":{"operationId":"BooksController_findOne","parameters":[{"name":"bookId","required":true,"in":"path","description":"ID of the book","schema":{"example":"2691fc73-e784-4e4c-a266-231f5992bfce","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookResponseDto"}}}},"404":{"description":"Book not found"},"500":{"description":"Internal server error"}},"summary":"Get a book by its ID","tags":["Books"]},"delete":{"operationId":"BooksController_remove","parameters":[{"name":"bookId","required":true,"in":"path","description":"ID of the book","schema":{"example":"2691fc73-e784-4e4c-a266-231f5992bfce","type":"string"}}],"responses":{"204":{"description":"Book deleted"},"404":{"description":"Book not found"},"500":{"description":"Internal server error"}},"summary":"Delete a book by its ID","tags":["Books"]}},"/auth/sign-in":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}},"500":{"description":"Internal server error"}},"summary":"Log into the app","tags":["Auth"]}},"/auth/sign-up":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpResponseDto"}}}},"500":{"description":"Internal server error"}},"summary":"Create an account","tags":["Auth"]}},"/users/me":{"get":{"description":"Get informations about the logged in user","operationId":"UsersController_getCurrentUser","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CurrentUserResponseDto"}}}},"500":{"description":"Internal server error"}},"security":[{"bearer":[]}],"summary":"","tags":["Users"]}}},"info":{"title":"Reader","description":"The Reader API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ChapterDto":{"type":"object","properties":{"content":{"type":"string","example":"The chapter explains that Resistance affects everyone pursuing meaningful creative or personal work, regardless of talent, experience, or success, and that no one is exempt from struggling against it."},"number":{"type":"number","example":6},"title":{"type":"string","example":"Resistance Is Universal"}},"description":"DTO for a single book chapter.","required":["content","number","title"]},"CreateBookDto":{"type":"object","properties":{"chapters":{"type":"array","items":{"$ref":"#/components/schemas/ChapterDto"}},"summary":{"type":"string","example":"The War of Art argues that creative success depends on recognizing and relentlessly overcoming “Resistance”—the internal force of fear, procrastination, and self-doubt that prevents meaningful work."},"title":{"type":"string","example":"The War of Art"}},"description":"DTO used for creating a new book.","required":["chapters","summary","title"]},"BookResponseDto":{"type":"object","properties":{"id":{"type":"string","example":"6e426018-4251-4c2c-85db-23e8f9af19ae"},"summary":{"type":"string","example":"The War of Art argues that creative success depends on recognizing and relentlessly overcoming “Resistance”—the internal force of fear, procrastination, and self-doubt that prevents meaningful work."},"title":{"type":"string","example":"The War of Art"}},"description":"DTO representing a book entity returned in API responses, including all book details except the chapters array.","required":["id","summary","title"]},"SignInDto":{"type":"object","properties":{"password":{"type":"string","example":"9JrHmsFZK8J-a"},"username":{"type":"string","description":"Unique username","example":"john-doe"}},"description":"Login credentials","required":["password","username"]},"SignInResponseDto":{"type":"object","properties":{"access_token":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIn0.XYZ123"}},"description":"Response received after login","required":["access_token"]},"SignUpDto":{"type":"object","properties":{"password":{"type":"string","example":"9JrHmsFZK8J-a"},"username":{"type":"string","example":"john-doe"}},"description":"Sign-up credentials","required":["password","username"]},"SignUpResponseDto":{"type":"object","properties":{"userId":{"type":"string","example":"6e426018-4251-4c2c-85db-23e8f9af19ae"},"username":{"type":"string","example":"john-doe"}},"description":"Response receive after creating an account","required":["userId","username"]},"CurrentUserResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"ID of the user","example":"2691fc73-e784-4e4c-a266-231f5992bfce"}},"required":["id"]}}}}